+++ core/domain/interface/content/content.go
@@ -41,7 +41,7 @@ const (
 type (
 	IContent interface {
 		// 获取聚合根编号
-		GetAggregateRootId() int32
+		GetAggregateRootId() int
 		// 文章服务
 		ArticleManager() IArticleManager
 		// 创建页面
@@ -67,7 +67,7 @@ type (
 
 	IContentRepo interface {
 		// 获取内容
-		GetContent(userId int32) IContent
+		GetContent(userId int64) IContent
 		// 根据编号获取页面
 		GetPageById(userId, id int32) *Page
 		// 根据标识获取页面
@@ -100,7 +100,7 @@ type (
 		// 编号
 		Id int32 `db:"id" pk:"yes" auto:"yes"`
 		// 商户编号
-		UserId int32 `db:"user_id"`
+		UserId int64 `db:"user_id"`
 		// 标题
 		Title string `db:"title"`
 		// 字符标识
