+++ core/domain/interface/merchant/merchant.go
@@ -84,7 +84,7 @@ type (
 	// 账户
 	IAccount interface {
 		// 获取领域对象编号
-		GetDomainId() int32
+		GetDomainId() int64
 		// 获取账户值
 		GetValue() *Account
 		// 保存
@@ -275,7 +275,7 @@ type (
 	// 商户账户表
 	Account struct {
 		// 商户编号
-		MchId int32 `db:"mch_id" pk:"yes"`
+		MchId int64 `db:"mch_id" pk:"yes"`
 		// 余额
 		Balance float32 `db:"balance"`
 		// 冻结金额
@@ -301,7 +301,7 @@ type (
 		// 编号
 		Id int32 `db:"id" pk:"yes" auto:"yes"`
 		// 商户编号
-		MchId int32 `db:"mch_id"`
+		MchId int64 `db:"mch_id"`
 		// 日志类型
 		Kind int `db:"kind"`
 		// 标题
@@ -325,7 +325,7 @@ type (
 		// 编号
 		Id int32 `db:"id" pk:"yes" auto:"yes"`
 		// 商户编号
-		MchId int32 `db:"mch_id"`
+		MchId int64 `db:"mch_id"`
 		// 新增订单数量
 		OrderNumber int `db:"order_number"`
 		// 订单额
